backend/requirements.txt
python-dotenv
openai-whisper==20240930
openai==1.57.2
ffmpeg
fastapi[standard]
python-docx
markdown
pytest
pymongo
dnspython
pyjwt
passlib[bcrypt]
yt-dlp
python-multipart
beautifulsoup4
pypdf
googletrans
pytest-asyncio
pytest-mock
mongomock
pre-commit
black
isort
mypy
types-Markdown
pytest-cov